Added Shell sort

+++ src/InsertionSort.cc
@@ -1,7 +1,9 @@
 
+#include<iostream>
+#include "myheaders.h"
 
 
-void Insert(long array[],long size,long start,long stop){
+void Insert(long array[],int size,int start,int stop){
 	long Needed=array[stop];
 
 	for(int i=stop;i>=start;--i){
@@ -13,8 +15,8 @@ void Insert(long array[],long size,long start,long stop){
 }
 
 
-void InsertionSort(long array[],long size){
-	for(long i=0;i<size;i++){
+void InsertionSort(long array[],int size){
+	for(int i=0;i<size;i++){
 		int j=0;
 		while(array[j]<array[i]){
 			j++;
